Move some resize helpers to separate file closer to the component
